requirements.txt
alembic==1.11.1
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
executing==1.2.0
Flask
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
python-dotenv
six
SQLAlchemy
stack-data
traitlets
typing_extensions
Wand
wcwidth
Werkzeug
WTForms
WTForms-SQLAlchemy
